TRAVELMIND/requirements.txt
langchain
langgraph
langchain-openai
langchain-community
langchain_openrouter
langchain-google-genai
requests
chromadb
sentence-transformers

flask

ipykernel

tinydb

python-dotenv
openai

tiktoken
pydantic
